V1  - multiple host support  - host registration  - reverse tcp connection to host for control  - config file for supported commands on tcp  listener  - bridge to facilitate connection between  mediamtx dynamically generated endpoints and  dashboard  - authentication, using semi-secure methods
--- tdlr ---
Host queries `:4000/register` endpoint on server
passes their tcp control ports, desired name and
registration token.

If OK , server opens a new udp port and sends
the port to the device, the device starts main app
to send traffic over to that udp port.

Traffic is recognized and mediamtx's opens stream,
using ffmpeg re-encodes the stream, and re-directs
it to a known endpoint (based on the cam name),
given on registration.

The dashboard queries the list on relaod and
creates an instance of the camera feed, with
controls.

On click of send, values from the frontend are
transformed into the correct format expected by
the tcp listener, and are sent by the server
through /cmd endpoint to the devices's external
ips, and their specified ports.

If device is behind a private network,
most likely port forwarding will be required.